requirements.txt
anyio==3.7.1
asyncmy==0.2.9
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==2.1.5
pydantic==1.10.19
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.12
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0.post1
